+++ trunk/src/OpenMesh/Core/Utils/RandomNumberGenerator.hh
@@ -56,6 +56,8 @@
 
 
 #include <OpenMesh/Core/System/config.h>
+#include <cstdlib>
+
 
 
 //== NAMESPACES ===============================================================
@@ -68,6 +70,7 @@ namespace OpenMesh {
 
 
 /**  Generate a random number between 0.0 and 1.0 with a guaranteed resolution
+ *   ( Number of possible values )
  *
  * Especially useful on windows, as there MAX_RAND is often only 32k which is
  * not enough resolution for a lot of applications
@@ -80,20 +83,20 @@ public:
   *
   * @param _resolution specifies the desired resolution for the random number generated
   */
-  RandomNumberGenerator(const double _resolution);
+  RandomNumberGenerator(const size_t _resolution);
 
   /// returns a random double between 0.0 and 1.0 with a guaranteed resolution
   double getRand() const;
 
-  double resolution() const;
+  size_t resolution() const;
 
 private:
 
   /// desired resolution
-  const double resolution_;
+  const size_t resolution_;
 
   /// number of "blocks" of RAND_MAX that make up the desired _resolution
-  unsigned int iterations_;
+  size_t iterations_;
 
   /// maximum random number generated, which is used for normalization
   double maxNum_;
